requirements.txt
torch
transformers
Pillow
opencv-python
scikit-learn
streamlit
python-dotenv
